centralized toggle functions #54
