+++ location/views/Profile.py
@@ -54,48 +54,8 @@ class ProfileView(UpdateView):
     return super().form_valid(form)
   
 
-''' Toggle Favorite '''
-class ToggleFavoriteLocation(UpdateView):
-  model = Location
-  fields = ['slug']
 
-  def get(self, request, *args, **kwargs):
-    location = Location.objects.get(slug=kwargs['slug'])
-    if hasattr(self.request.user, 'profile'):
-      profile = self.request.user.profile
-    else:
-      profile = Profile.objects.create(user=self.request.user)
-      messages.add_message(self.request, messages.INFO, f"Created profile for { self.request.user.get_full_name() }")
-    ''' If Location is in Favorites, remove it '''
-    if location in profile.favorite.all():
-      profile.favorite.remove(location)
-      messages.add_message(self.request, messages.SUCCESS, f"{ _('Removed') } { location } { _('from favorites') }.")
-    else:
-      profile.favorite.add(location)
-      messages.add_message(self.request, messages.SUCCESS, f"{ _('Added') } { location } { _('to favorites') }.")
-    return redirect('location:location', location.slug)
 
-''' Toggle Family Member '''
-class ToggleFamilyMember(UpdateView):
-  model = User
-  fields = ['id']
-
-  def get(self, request, *args, **kwargs):
-    family_member = User.objects.get(id=kwargs['id'])
-    if hasattr(self.request.user, 'profile'):
-      profile = self.request.user.profile
-    else:
-      profile = Profile.objects.create(user=self.request.user)
-      messages.add_message(self.request, messages.INFO, f"Created profile for { self.request.user.get_full_name() }")
-    ''' If User is already a family member '''
-    if family_member in profile.family.all():
-      profile.family.remove(family_member)
-      messages.add_message(self.request, messages.SUCCESS, f"{ _('Removed') } { family_member.get_full_name() } { _('from family') }.")
-    else:
-      profile.family.add(family_member)
-      messages.add_message(self.request, messages.SUCCESS, f"{ _('Added') } { family_member.get_full_name() } { _('to family') }.")
-    return redirect('location:profile')
-  
 
 
 ''' Sign up '''
