+++ location/views/__init__.py
@@ -5,3 +5,5 @@ from .Tag import *
 from .List import *
 from .Profile import *
 from .Chain import *
+
+from .Toggle import *
